Remove duplicate parse_rinex_nav_file function (#159)
* replaces poetry by uv

* oups

* format

* update readme

* remove duplicate nav parse function

* fix

* fix

* merged helpers.py into util.py

* hmm
