+++ src/prx/main.py
@@ -7,18 +7,21 @@ import georinex
 import pandas as pd
 import numpy as np
 import git
+import prx.util
 
-from prx import atmospheric_corrections as atmo
+from prx import atmospheric_corrections as atmo, util
 from prx.constants import carrier_frequencies_hz
-from prx.helpers import parse_rinex_file
+from prx.util import parse_rinex_obs_file
+from prx.util import is_rinex_3_obs_file, is_rinex_3_nav_file
 from prx.rinex_nav import nav_file_discovery
-from prx import constants, helpers, converters, user
+from prx import constants, converters, user
 from prx.rinex_nav import evaluate as rinex_evaluate
+from prx.rinex_nav.evaluate import parse_rinex_nav_file
 
-log = helpers.get_logger(__name__)
+log = util.get_logger(__name__)
 
 
-@helpers.timeit
+@prx.util.timeit
 def write_prx_file(
     prx_header: dict,
     prx_records: pd.DataFrame,
@@ -198,7 +201,7 @@ def build_metadata(input_files):
     prx_metadata["input_files"] = [
         {
             "name": file.name,
-            "murmur3_hash": helpers.hash_of_file_content(file),
+            "murmur3_hash": util.hash_of_file_content(file),
         }
         for file in files
     ]
@@ -221,11 +224,21 @@ def check_assumptions(
     )
 
 
+def parse_rinex_nav_or_obs_file(rinex_file_path: Path):
+    if is_rinex_3_obs_file(rinex_file_path):
+        return parse_rinex_obs_file(rinex_file_path)
+    elif is_rinex_3_nav_file(rinex_file_path):
+        return parse_rinex_nav_file(rinex_file_path)
+    assert False, (
+        f"File {rinex_file_path} appears to be neither RINEX 3 OBS nor NAV file."
+    )
+
+
 def warm_up_parser_cache(rinex_files):
-    _ = [parse_rinex_file(file) for file in rinex_files]
+    _ = [parse_rinex_nav_or_obs_file(file) for file in rinex_files]
 
 
-@helpers.timeit
+@prx.util.timeit
 def build_records(
     rinex_3_obs_file,
     rinex_3_ephemerides_files,
@@ -236,7 +249,7 @@ def build_records(
         approximate_receiver_ecef_position_m
     )
     check_assumptions(rinex_3_obs_file)
-    flat_obs = helpers.parse_rinex_obs_file(rinex_3_obs_file)
+    flat_obs = util.parse_rinex_obs_file(rinex_3_obs_file)
 
     flat_obs.time = pd.to_datetime(flat_obs.time, format="%Y-%m-%dT%H:%M:%S")
     flat_obs.obs_value = flat_obs.obs_value.astype(float)
@@ -357,17 +370,15 @@ def build_records(
         how="left",
     )
     # Compute anything else that is satellite-specific
-    sat_states["relativistic_clock_effect_m"] = (
-        helpers.compute_relativistic_clock_effect(
-            sat_states[["sat_pos_x_m", "sat_pos_y_m", "sat_pos_z_m"]].to_numpy(),
-            sat_states[["sat_vel_x_mps", "sat_vel_y_mps", "sat_vel_z_mps"]].to_numpy(),
-        )
+    sat_states["relativistic_clock_effect_m"] = util.compute_relativistic_clock_effect(
+        sat_states[["sat_pos_x_m", "sat_pos_y_m", "sat_pos_z_m"]].to_numpy(),
+        sat_states[["sat_vel_x_mps", "sat_vel_y_mps", "sat_vel_z_mps"]].to_numpy(),
     )
-    sat_states["sagnac_effect_m"] = helpers.compute_sagnac_effect(
+    sat_states["sagnac_effect_m"] = util.compute_sagnac_effect(
         sat_states[["sat_pos_x_m", "sat_pos_y_m", "sat_pos_z_m"]].to_numpy(),
         approximate_receiver_ecef_position_m,
     )
-    [latitude_user_rad, longitude_user_rad, height_user_m] = helpers.ecef_2_geodetic(
+    [latitude_user_rad, longitude_user_rad, height_user_m] = util.ecef_2_geodetic(
         approximate_receiver_ecef_position_m
     )
     days_of_year = np.array(
@@ -378,7 +389,7 @@ def build_records(
     (
         sat_states["elevation_rad"],
         sat_states["azimuth_rad"],
-    ) = helpers.compute_satellite_elevation_and_azimuth(
+    ) = util.compute_satellite_elevation_and_azimuth(
         sat_states[["sat_pos_x_m", "sat_pos_y_m", "sat_pos_z_m"]].to_numpy(),
         approximate_receiver_ecef_position_m,
     )
@@ -475,7 +486,7 @@ def build_records(
             time_of_emission_weeksecond_isagpst = (
                 flat_obs.loc[mask]
                 .apply(
-                    lambda row: helpers.timedelta_2_weeks_and_seconds(
+                    lambda row: util.timedelta_2_weeks_and_seconds(
                         row.time_of_emission_isagpst
                         - constants.system_time_scale_rinex_utc_epoch["GPST"]
                     )[1],
@@ -513,7 +524,7 @@ def build_records(
     return flat_obs
 
 
-@helpers.timeit
+@prx.util.timeit
 def process(observation_file_path: Path, output_format="csv"):
     t0 = pd.Timestamp.now()
     # We expect a Path, but might get a string here:
@@ -522,7 +533,7 @@ def process(observation_file_path: Path, output_format="csv"):
         f"Starting processing {observation_file_path.name} (full path {observation_file_path})"
     )
     rinex_3_obs_file = converters.anything_to_rinex_3(observation_file_path)
-    rinex_3_obs_file = helpers.repair_with_gfzrnx(rinex_3_obs_file)
+    rinex_3_obs_file = prx.util.repair_with_gfzrnx(rinex_3_obs_file)
     prx_file = rinex_3_obs_file.with_suffix("")
     aux_files = nav_file_discovery.discover_or_download_auxiliary_files(
         rinex_3_obs_file
@@ -531,7 +542,7 @@ def process(observation_file_path: Path, output_format="csv"):
         {"obs_file": rinex_3_obs_file, "nav_file": aux_files["broadcast_ephemerides"]}
     )
     metadata["processing_start_time"] = t0
-    helpers.repair_with_gfzrnx(rinex_3_obs_file)
+    prx.util.repair_with_gfzrnx(rinex_3_obs_file)
     records = build_records(
         rinex_3_obs_file,
         aux_files["broadcast_ephemerides"],
