+++ src/prx/rinex_nav/test/test_nav_file_discovery.py
@@ -5,9 +5,11 @@ import shutil
 import pandas as pd
 import pytest
 import subprocess
-import georinex
+
+from prx import util
 from prx.rinex_nav import nav_file_discovery as aux
-from prx import helpers, converters
+import georinex
+from prx import converters
 
 
 @pytest.fixture
@@ -22,12 +24,12 @@ def set_up_test():
     test_nav_file = test_directory.joinpath("BRDC00IGS_R_20230010000_01D_MN.rnx.zip")
 
     shutil.copy(
-        helpers.prx_repository_root()
+        util.prx_repository_root()
         / f"src/prx/test/datasets/TLSE_2023001/{test_obs_file.name}",
         test_obs_file,
     )
     shutil.copy(
-        helpers.prx_repository_root()
+        util.prx_repository_root()
         / f"src/prx/test/datasets/TLSE_2023001/{test_nav_file.name}",
         test_nav_file,
     )
@@ -53,7 +55,7 @@ def test_download_remote_ephemeris_files(set_up_test):
 def test_command_line_call(set_up_test):
     test_file = set_up_test["test_obs_file"]
     aux_file_script_path = (
-        helpers.prx_repository_root() / "src/prx/rinex_nav/nav_file_discovery.py"
+        util.prx_repository_root() / "src/prx/rinex_nav/nav_file_discovery.py"
     )
 
     command = f"python {aux_file_script_path} --observation_file_path {test_file}"
@@ -97,12 +99,10 @@ def test_use_local_nav(set_up_test):
     if local_database_file is None:  # if BRDC NAV file is not in local database
         # download BRDC Nav file
         header = georinex.rinexheader(set_up_test["test_obs_file"])
-        t_start = helpers.rinex_header_time_string_2_timestamp_ns(
+        t_start = util.rinex_header_time_string_2_timestamp_ns(
             header["TIME OF FIRST OBS"]
         ) - pd.Timedelta(200, unit="milliseconds")
-        t_end = helpers.rinex_header_time_string_2_timestamp_ns(
-            header["TIME OF LAST OBS"]
-        )
+        t_end = util.rinex_header_time_string_2_timestamp_ns(header["TIME OF LAST OBS"])
         aux.update_local_database(t_start, t_end)
     assert local_database_file.exists()
