+++ src/prx/sp3/evaluate.py
@@ -6,14 +6,13 @@ import georinex
 from pathlib import Path
 import matplotlib.pyplot as plt
 
-from prx import helpers
-from prx import constants
+from prx import constants, util
 
-log = helpers.get_logger(__name__)
+log = util.get_logger(__name__)
 
 
 def parse_sp3_file(file_path: Path):
-    @helpers.disk_cache.cache(ignore=["file_path"])
+    @util.disk_cache.cache(ignore=["file_path"])
     def cached_load(file_path: Path, file_hash: str):
         log.info(f"Parsing {file_path} (hash {file_hash}) ...")
         parsed = georinex.load(file_path)
@@ -34,7 +33,7 @@ def parse_sp3_file(file_path: Path):
                 df.rename(columns={col: col.replace("_x", "")}, inplace=True)
         # Convert timestamps to seconds since GPST epoch
         df["gpst_s"] = (df["time"] - constants.cGpstUtcEpoch).apply(
-            helpers.timedelta_2_seconds
+            util.timedelta_2_seconds
         )
         df.drop(columns=["time"], inplace=True)
         df["sat_clock_offset_m"] = (
@@ -69,7 +68,7 @@ def parse_sp3_file(file_path: Path):
         df.insert(0, "gpst_s", df.pop("gpst_s"))
         return df
 
-    file_content_hash = helpers.hash_of_file_content(file_path)
+    file_content_hash = util.hash_of_file_content(file_path)
     return cached_load(file_path, file_content_hash)
 
 
@@ -156,7 +155,7 @@ def compute(sp3_file_path, query):
         if len(samples.index) > 0:
             sat_pv = interpolate(
                 samples,
-                helpers.timedelta_2_seconds(
+                util.timedelta_2_seconds(
                     row["query_time_isagpst"] - constants.cGpstUtcEpoch
                 ),
             )
